+++ test-resources-client/src/main/java/io/micronaut/testresources/client/TestResourcesClientPropertyExpressionResolver.java
@@ -17,7 +17,6 @@ package io.micronaut.testresources.client;
 
 import io.micronaut.context.env.Environment;
 import io.micronaut.context.env.PropertyExpressionResolver;
-import io.micronaut.core.annotation.Nullable;
 import io.micronaut.core.convert.ConversionService;
 import io.micronaut.core.value.PropertyResolver;
 import io.micronaut.testresources.core.LazyTestResourcesExpressionResolver;
@@ -25,9 +24,6 @@ import io.micronaut.testresources.core.TestResourcesResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
@@ -48,40 +44,6 @@ public class TestResourcesClientPropertyExpressionResolver extends LazyTestResou
 
     private static TestResourcesClient createClient(Environment env) {
         return TestResourcesClientFactory.findByConvention().orElse(NoOpClient.INSTANCE);
-    }
-
-    private static class NoOpClient implements TestResourcesClient {
-        static final TestResourcesClient INSTANCE = new NoOpClient();
-
-        @Override
-        public List<String> getResolvableProperties(Map<String, Collection<String>> propertyEntries, Map<String, Object> testResourcesConfig) {
-            return Collections.emptyList();
-        }
-
-        @Override
-        public Optional<String> resolve(String name, Map<String, Object> properties, Map<String, Object> testResourcesConfig) {
-            return Optional.empty();
-        }
-
-        @Override
-        public List<String> getRequiredProperties(String expression) {
-            return Collections.emptyList();
-        }
-
-        @Override
-        public List<String> getRequiredPropertyEntries() {
-            return Collections.emptyList();
-        }
-
-        @Override
-        public boolean closeAll() {
-            return true;
-        }
-
-        @Override
-        public boolean closeScope(@Nullable String id) {
-            return true;
-        }
     }
 
     private static class DelegateResolver implements PropertyExpressionResolver, AutoCloseable {
