+++ test-resources-extensions/test-resources-extensions-junit-platform/build.gradle
@@ -59,6 +59,7 @@ dependencies {
     testRuntimeOnly(libs.junit.jupiter.engine)
     testRuntimeOnly(mn.micronaut.context)
     testRuntimeOnly(mn.snakeyaml)
+    testRuntimeOnly(mnSerde.micronaut.serde.jackson)
     kspKoTest(mn.micronaut.inject.kotlin)
 }
